requirements.txt
pandas
numpy
matplotlib
scipy
scikit-learn
plotly
Pillow
kaleido
requests
PyGithub
yfinance
pandas-ta
streamlit
stocktrends